Set server_time = Double to fix parsing errors.

+++ ChatRoom/Message.swift
@@ -8,11 +8,10 @@
 
 import Foundation
 
-// Message Class
 struct Message: Codable {
     var msg: String
     var client_time:Double
-    var server_time:Int?
+    var server_time:Double?
     
     init(msg: String, client_time: Double) {
         self.msg = msg
@@ -20,9 +19,11 @@ struct Message: Codable {
         self.server_time = nil
     }
     
-    init(msg: String, client_time: Double, server_time: Int) {
+    init(msg: String, client_time: Double, server_time: Double) {
         self.msg = msg
         self.client_time = client_time
         self.server_time = server_time
     }
 }
+
+
